Fix resources to load from domain root
I don't know if this is best, but it's what is done on some example sites.  It
would be nice to use some kind of {{page.location}} or something to make
relative paths so the site could be deployed under a non-root location but I
don't know if that exists and don't want to find out right now.